+++ src/System.Private.CoreLib/shared/System/MissingMemberException.cs
@@ -17,7 +17,7 @@ namespace System
 {
     [Serializable]
     [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
-    public class MissingMemberException : MemberAccessException
+    public partial class MissingMemberException : MemberAccessException
     {
         public MissingMemberException()
             : base(SR.Arg_MissingMemberException)
@@ -57,23 +57,6 @@ namespace System
             info.AddValue("MMClassName", ClassName, typeof(string));
             info.AddValue("MMMemberName", MemberName, typeof(string));
             info.AddValue("MMSignature", Signature, typeof(byte[]));
-        }
-
-        public override string Message
-        {
-            get
-            {
-                return ClassName == null ? base.Message : SR.Format(SR.MissingMember_Name, ClassName + "." + MemberName + (Signature != null ? " " + FormatSignature(Signature) : string.Empty));
-            }
-        }
-
-        internal static string FormatSignature(byte[] signature)
-        {
-            // This is not the correct implementation, however, it's probably not worth the time to port given that 
-            //  (1) it's for a diagnostic
-            //  (2) Signature is non-null when this exception is created from the native runtime. Which we don't do in .Net Native.
-            //  (3) Only other time the signature is non-null is if this exception object is deserialized from a persisted blob from an older runtime.
-            return string.Empty;
         }
 
         // If ClassName != null, GetMessage will construct on the fly using it
