+++ src/System.Private.CoreLib/src/System/MissingMemberException.CoreRT.cs
@@ -1,0 +1,37 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+/*=============================================================================
+**
+**
+**
+** Purpose: The exception class for versioning problems with DLLS.
+**
+**
+=============================================================================*/
+
+using System.Runtime.Serialization;
+
+namespace System
+{
+    public partial class MissingMemberException : MemberAccessException
+    {
+        public override string Message
+        {
+            get
+            {
+                return ClassName == null ? base.Message : SR.Format(SR.MissingMember_Name, ClassName + "." + MemberName + (Signature != null ? " " + FormatSignature(Signature) : string.Empty));
+            }
+        }
+
+        internal static string FormatSignature(byte[] signature)
+        {
+            // This is not the correct implementation, however, it's probably not worth the time to port given that 
+            //  (1) it's for a diagnostic
+            //  (2) Signature is non-null when this exception is created from the native runtime. Which we don't do in .Net Native.
+            //  (3) Only other time the signature is non-null is if this exception object is deserialized from a persisted blob from an older runtime.
+            return string.Empty;
+        }
+    }
+}
